refactor: validator at gnb
